Added the ability to remove panels at each layer in the Panel API
The purpose of this change was to enable perspective switching without
resource leaks in the templated panel system. That goal is not yet
achieved because the Templated panel system always uses a null Position
for child panels, and the only way to remove a child PanelDefinition is by
position (the new stuff I added does it by layer: Presenters remove
Presenters from themselves, Views remove Views, etc.). So it’s not currently
possible to remove a child the PanelDefinition.

I’m committing this now because it’s working properly for the classic panel
manager and the changeset is getting big.