Fixed UF-162: perspective geometry occasionally saves incorrectly
This fix is a bit of a band-aid for the underlying problem, which is that it
appears to be a race between the browser's layout engine assigning sizes to
elements and the JavaScript thread checking for element sizes, at least in
FireFox. A different but equally wonky band-aid would be to introduce small
delays before propagating the onResize() calls. Note that the onResize() was
already being propagated from a Sheduler.scheduleDeferred() and this wasn't
enough to achieve 100% reliability.

The best fix would be to implement the idea in UF-117. Then it would not
be necessary to drop size changes back to null.