Refix WAS shutdown issues after todays changes (see below)
Todays changes for fixing tests introduced unregistering
the systemFs from the PriorityDisposableRegistry when the
application scope is destroyed. Keeping the systemFs
available after the app scope is destroyed is however
the very reason we registered it. We have servlet listeners
that require the systemFs to be available during shutdown.

To come to a quick solution for this release and to keep
today's test fixes, the ConfigIOServiceProducer is now
holding on to a reference to systemFs once it successfuly
looked it up. So that it can hand it out during shutdown
even if it was unregistered.