[UF-444] Avoid concurrent Freemarker setup (#574)
The annotation processor sometimes failed when used
concurrently (e.g. in parallel Maven build). The reason
seems to be the fact that Freemarker's Configuration class
(or its dependencies) is not thread safe.

Synchronizing the Freemarker setup will likely make the init
sligthly slower, but since this is only build-time
(compile-time) processor it should not matter much.