Revert "proposal that not only removes Thread.sleep but also remove the (#628)"
This reverts commit 16639e0a7dc66a77bedfa47c1083a429199bbbc4.

The Observer class and mechanism removed by the is NOT just used by tests.
it's used to attach the "Message Console" to the indexing results too.
See org.kie.workbench.screens.workbench.backend.impl.DefaultApplicationScopedProducer.
This causes the downstream builds to fail.

The changes need to be split and only the indexing test fix proposed
again.