JBPM-5452: Case Modeller (#711)
This commit re-uses BPMN2 DefinitionSet for Case Management.
It is feature complete with tests for a proof of concept as
agreed to date. We need to provide Users with the ability
to open a file with either BPMN2 editor or Case Management
editor. That is to follow in a subsequent iteration.

Additional bugs and enhancements are logged against the
referenced JIRA, JBPM-5452.