GUVNOR-2854: Increase coverage for moving rows with PriorityOver (#425)
This PR increases coverage of moving rows with set PriorityOver and also
fixes two small issues.

The first issue was, that when te row with PriorityOver Row X was moved
to the position X, his PriorityOver wasn't cleared. It means that the
Row X with PriorityOver X was in the table.

The second fixed issue. When there were Rows: 1, 2, 3 with Priorities
None, 1, 1 and the Row 1 was moved to Row 3 then the priorites were None,
3, None. Now the priorities will be None, None, None as expected.