AF-1186: Versions history is lost after asset rename (#336)
Fixes `ListCommits` command to follow path name changes, and maps version records to correct paths of renamed files.

Note that there are issues with some editors when trying to view previous versions of moved files. One that I encountered is with the data modeller. That editor produces two commits when moving a file. One renames the Java class and the other moves the file. If you try checking out the commit that renames the Java class, the data modeller cannot load the class, presumably because the class name and file name do not match. We might want to make a JIRA issue to ensure that rename operations are committed atomically. //cc @wmedvede