[AF-1544]: Dot files are not indexed, so dot files condition added before queueing delete events (#466)
There was a NPE when trying to delete _src/main/.java_ from index, but those files are not directly indexed as they are. They are converted to _src/main/java_, so we can't delete a _.java_ that is not there. 
I added that check so now it will not try to delete dot files from index.