Drools 3347 - Implement undo/redo (#1007)
Scope of this PR is to actually implement the UNDO/REDO behavior.
In our context, UNOD/REDO means restoring the overall status at a given moment. In other words (especially for redo) it does not mean re-execution of a command. Whenever a new command is executed, the "redo" list is cleared, so it is not possible for ex to issue a "Insert new row" and then repeat it with "Redo"