[DROOLS-3417] Checking for zero header row count inside ScenarioSImulationGridHeaderUtilies. Importing jomarko modifications to test. (#1022)
That loop was caused by a change inside upper layer.
Before this change, by default ScenariogridModel had 1 header row, even when called by the default constructor. After the change, the default header row number is 0.
Inside ScenarioSimulationGridHeaderUtilities there is a method that uses that row number in a "while" loop for a subtraction, but since now that number is 0, that loop will never end.
The fix has been to check for "0" rows before that loop, eventually returning null.
Tests have been modified and a new one has been added.