KOGITO-7533 - Keyboard navigation though workflow states
Node Navigation using arrows is now possible. Use up, down, left and right keys to navigate.
Left and Right keys navigates through sibling nodes.
Up navigates to source nodes and down navigates to target nodes.
When using up or down it tries to go through the path it came from in cases where there are multiple targets or sources.
In VSCode the text editor still selects the line corresponding to the selected node, but it does not steal the focus from the canvas.