Issue #70 - Loading Unsaved asset in VSCode leads to an empty file
In Windows if the backup string is solver using Uri.parse(), an invalid URI is created.
The parse  method should be used to solve stringfied URIs. Which is not the case here.
The backupId is an actual filepath, so Uri.file() is the correct method.
The reason this error only happens in windows is because o the file system.
Unix file system based OSs are not affected by this.