kie-issues#41 DMN\BMPN Editors: The file is kept "dirty" after undoing one copy/paste operation
This PR closes: https://github.com/kiegroup/kie-issues/issues/41

Editor operations are implemented as commands, that are stored in a registry to have an option of undo an redo.
The situation is quite complicated for complex operations, that are internally implemented as multiple commands, but for the end user is visible as single operation.
For such, we must compose multiple commands into single one. In copy and paste scenario, we must go one level further and compose two already composed commands into single command.